Reduce the fadeout delay of the Loading indicator.
The Loading indicator was given a 1 second delay in order to not just
flash on and off on some requests, so that users can see that their
request had indeed started and finished.

1 second is a while, so this drops it down to 250ms. Testing with my
local server, it was on the screen just long enough to show the request
started and completed, but short enough to feel like the request
finished quickly.

This should give the perception of a large speed boost.