Support adding/removing users with RB.ReviewGroup.
RB.ReviewGroup can now add users to the group or remove users from the
group through the new addUser/removeUser functions.

This works much in the same way as watched object tracking. We have a
proxy object, GroupMember, that we can call save() or destroy() on, and
it will invoke the POST or DELETE request with the data needed for
manipulating that membership.