Add a darker search icon to the spritesheet.
This adds a darker version of the search icon, intended for lighter
backgrounds. It's otherwise of the same style.

For consistency, I placed this as the next icon after the lighter search
icon, shifting over some other icons.

I also shifted the light icon just a bit to better pixel-align it.