Add a Backbone collection for filtering other collections.
This adds RB.FilteredCollection, which presents a filtered view of
another collection. FilteredCollection takes a dictionary of filters,
which map field names to values. For string fields, the prefix will be
compared, and for other values, direct comparison is used. This can be
expanded as needed down the road to do more complex matches.

When the main collection updates, the FilteredCollection will update as
well, according to the filter.