Add some infrastructure for JavaScript-controlled config pages.
This is a port of some of the code in RBCommons used to handle the
admin and My Account pages. It provides some CSS and JavaScript to
display lists of items with optional actions for config pages.

This is, for now, intended to be a subset of the infrastructure used in
RBCommons, and is just enough for the new My Account page. Going
forward, we'll likely extend this for new bits of UI and things like
better dynamic loading of data.