Filter out replies when calculating issue counts.
Issue counts on review requests were going up when replies were made to
comments with issues. This happened because we looked for all comments,
and not just top-level comments, when calculating issue counts and when
deciding whether to increase issue counts.

We now limit these only to top-level reviews/comments.

I also noticed that we were bumping the last review activity timestamp
even when a review was still in draft mode. That's been fixed so that
the value is only updated when the review is public.