Improve the Markdown escaping logic.
This improves our Markdown escaping to better handle parens,
underscores, and asterisks.

The common cases of showing function names with parens and underscores
should no longer result in a bunch of escaping. We now do a better job
of escaping only when they're going to cause formatting changes.

Since some of the regexes were getting a bit longer, I've also
repositioned the comments in the regexes to appear before the regex
lines.