Improve line number reference URLs for the text review UI.
For the text review UI, you can reference a line number by
appending #source/line123 to the URL. This improves it a bit by allowing
just #line123.

Since Backbone.Router isn't very smart when it comes to
pattern-matching, we have no way of specifying a mode just for line
numbers. The greediest entry seems to win, and that'd be the :viewMode
one. So, we need to transform in the router handler.