Add render_value to various PasswordInput fields.
For password inputs that are part of settings (LDAP auth, mail host, and
repository), the password wasn't pre-populated into the form, which meant that
if you changed said settings, you'd need to re-enter the password. This is
highly annoying. I've added `render_value=True` to the affected PasswordInput
ctors.

Testing done: Checked that saving a password was "sticky".

Fixes bugs 3037, 3167.