Issue 3200: "Depends on" field now uses strikethrough for submitted dependencies.
The "Depends on" field is used to indicate when one change depends on another.
It would be very useful if it could react when those dependencies are marked as
submitted by setting a strikethrough on the text. That way, users don't have to
click through to see whether or not the dependent change has been committed
yet.

Testing Done:
Manual testing: marked a dependency submitted, observed it show up with a
linethrough in the "Depends On" list for another review. All tests in djblets
and reviewboard pass.

Bugs closed: 3200

Reviewed at https://reviews.reviewboard.org/r/5348/