Clearcase often fails in multi-site mode because oids are not yet available on server
When people are submitting a review request, the oid of the old file version
and new file version are sent and used to retrieve the filename. However, the
new filename is usually not yet synced. Since the filename is not mandatory, we
can catch this exception and try to retrieve path from client filename and vob.

Testing Done:
Successfully created a review request before the replica was sent.

Bugs closed: 3183

Reviewed at https://reviews.reviewboard.org/r/5185/