Fix possible duplication issues with ReviewRequestField registration.
ReviewRequestFieldSet has a global list for storing any registered
fields. If a subclass didn't override that list, then it would share the
list of fields with any other subclass that didn't override it, leading
to problems.

We now use the pattern that the My Account page uses, where we default
the list to None, and then we specifically set it to an empty list on
the subclass during fieldset registration, if it's unset.