Update much of the documentation for the Extensions Overview page.
This rewrites much of the docs to simplify some of the language, clarify
some points, to modernize how things work, and to link to more useful
Django docs.

The general structure is still left the same. My next change is going to
split up the page into more dedicated topic pages.