Redo the review request ID column to fix breakages.
The new registration system for datagrid columns in Djblets was
incompatible with the way that the review_id column works for the
dashboard. This column was being overwritten in __init__, which meant
that it skipped registration. The original column had no way of grabbing
the right field itself, resulting in a breakage.

The hack is now gone, and a dedicated column has been created for the
review request ID. I've also shortened the label and made it more clear
this is about review request IDs, not review IDs.