Switch the "check updates required" middleware to use process_view.
The introduction of apply_to for media bundles was interfering with the updates
required middleware. In particular, what was happening was that we tried to
render a template from the process_request method of a middleware, which
happens before HttpRequest.resolver_match is set. This is easily fixed by using
process_view instead, which is probably more correct for this middleware anway.

This problem exhibited itself when running the
testManualUpdatesRequiredBadUpload unit test. I've checked all our other
middlewares to make sure that none of them are trying to render responses from
process_request.

Testing done: Ran unit tests.

Reviewed at https://reviews.reviewboard.org/r/5389/