Fix a regression involving the new visible check in the repository page.
The new visible check that was recently added broke RBCommons, because
of the lack of a displayed 'visible' checkbox. Since the new repository
work for Review Board may also be impacted by this in the future, I'm
fixing it here. We now try to get the value, and fall back on it being
visible.