Move the datagrids code out into reviewboard.datagrids.
The datagrids code is a hefty portion of reviewboard.reviews, and is
going to be heftier soon. Especially considering it was bundled into one
big file. It also felt strange to have extensions import from
reviewboard.reviews.datagrids for things that weren't related to review
requests.

This introduces reviewboard.datagrids, which contains all the columns,
datagrids, URL mappings, and views.

The datagrid and column code have been untouched, with a couple
exceptions. I rearranged the columns to be in alphabetical order,
renamed StarColumn to BaseStarColumn, added a missing localization for
the label for NewUpdatesColumn, and replaced double quotes with single
quotes. No logic has changed anywhere.