Move the user page to the new sidebar classes.
The user page has been updated to use the new sidebar code for the
profile and group information. This improves the rendering of the
groups, makes the page easier to maintain, and also improves
extensibility.

With this, the existing hooks are now useless, and will be removed in a
future patch.

I've also cleaned up how the two different modes for the sidebar work
(scrolling vs. fit-to-content). A "scrollable" CSS class on the sidebar
puts the sidebar into scrolling mode. The default is now fit-to-content.