Don't auto-collapse review boxes that have open issues or drafts.
It was easy to lose open issues after posting a change, since the boxes
would collapse. It was also annoying to have to reopen old reviews that
you were still commenting on (though at least the draft banner was
visible).

We now remove the collapsed state on boxes if they have any open issues
listed (and the viewing user is the author of the review request),
or if there's any draft reviews, making it easier to see the reviews
that still need attention.