Modernize our Getting Started guide.
This adds some useful missing information to Getting Started and removes
some old information.

The Windows steps are now gone, since we're not supporting development
on Windows without a VM.

The post-review instructions have been rewritten to reference rbt post,
and to encourage usage of -g and -u.

There's also some info on how to update your database when there are
schema changes.