Fix population of incoming request count on groups for the sidebar.
Due to the way we were loading the data for groups on the sidebar, we
weren't always showing counts for groups. If the incoming request count
wasn't previously calculated (or was reset, such as on upgrade), we
simply wouldn't show a counter until that group was next loaded by the
user in some way.

We now load the groups in a normal way, instead of using values().