Modernize the Review UI extension documentation.
Much of the Review UI documentation was old, based on what we had when
we wrote 1.7.

This updates it to make use of the static bundles and JavaScript
classes. It's now shown to be a lot simpler to build.

It still leaves out much of the Review UI functionality, like custom
templates or diffing, but it's at least less misleading.