Fix for an IntegrityError when trying to load the dashboard
When trying to load the dashboard using the datagrids module, the profile was
not passed as a parameter to the site_profile property.

Testing Done:
Tried to open the dashboard in the logged in as well as the logged out state to
check if the error appears again. Also registered as a new user and saw whether
the issue occurs.

Reviewed at https://reviews.reviewboard.org/r/5435/