Don't set width: 100% for all images in the review request.
A recent fix for thumbnail overflows resulted in every image on a review
request box being set to a 100% width. This affected the mimetype icons
for file attachments and any images in the text areas.

The other fix that was part of the commit introducing this appears to be
sufficient for the original overflow issue.