Don't show the expand icon if a header isn't expandable.
If multiple sections of collapsed context exist within a single function, the
headers listed in the collapsed sections will be the same function name, but
only the first one is expandable. Unfortunately, the "+" icon was still shown,
despite it not having an <a> tag around it.

Testing done:
Loaded a diff where I noticed this issue. Saw that the function name was still
shown, but the expand icon was gone.

Reviewed at https://reviews.reviewboard.org/r/5476/