Add a "Diff Size" column to the dashboard.
This change adds a new column to our Review request datagrids that shows an
approximate size for the diff (added and removed lines). It doesn't correspond
very directly to what users will see in the diff viewer, but it should be good
enough to give a general sense of how big the patch is.

Testing done:
Added the column to my dashboard when running a dump of the
reviews.reviewboard.org database. Saw nice diff sizes for all changes that
included a diff, and nothing for attachment-only review requests.

Fixes bug 48.

Reviewed at https://reviews.reviewboard.org/r/5478/