Improve the style and maintainability of the new Markdown links.
This polishes up the Markdown links for the various edit fields. They're
now blue and don't have an underline, and they're also much shorter.
Particularly for the comment dialog, where it was taking up valuable
space and making it harder to drag the window.

For the comment dialog, the text is now just "Markdown." For the other
fields, where we have room, it's "Markdown Reference."

This also cleans up the code a bit to put the links in the templates
instead of injecting them into the DOM elsewhere. For comment dialogs,
it's in the template now (and no longer displays when showing the
"you must log in" or "it's a draft!" messages). For the inlineEditors,
it's now MarkdownEditorView that sets it up and not each individual
instance.

The root of the user manual is now a main JavaScript constant instead of
repeated throughout the codebase, making it easier to maintain. Each of
these links reference it.

There's also some fixes for the links in the comment dialog, which
weren't pointing to valid addresses.