Trivial fix for publishing with comma-editable fields and None values.
Comma-editable fields were, by default, wrapping the old and new values
in a set() when checking if the values had changed, in order to do
order-independent comparisons.

However, if the value was None, instead of an empty list, they would
error out. A None value may be perfectly legal for that field, so we
need to allow for this.

This change fixes this by converting None to an empty list when building
the sets.