Fix move flags bumping over line numbers.
In the diff viewer, the flagpole for moved-line flags would cause a
misalignment of the line numbers. The diff viewer style had some code to
properly cope with the added border that move flags introduced, but it was only
active for replace blocks (because we used to put the flags for delete/add
blocks) on the other side of the text column. I've made it so this rule works
for all ".moved-row th" elements.

Testing done:
Loaded a diff that had some moved code. Saw that the line numbers were all
properly aligned.

Reviewed at https://reviews.reviewboard.org/r/5500/