Add a Sphinx extension to copy over Retina images.
Sphinx only actually copies over images that are used in documents. This
means that while we can dump @2x copies of images into _static, they
won't be used.

This extension hooks into the environment generation and html page
collection events, looking for @2x equivalents for all images that are
in use. It then adds them to the list of images set to be copied into
_images.