Handle LDAP users who don't have email addresses configured.
When a given field isn't defined for a user, the result from ldap searches will
not include that key at all. This means that in the case of users who don't
have an email address set, they won't be able to log in to Review Board.

This change handles KeyError in get_or_create_user when using a special email
attribute and logs the error instead of failing.

Fixes bug 3257.

Reviewed at https://reviews.reviewboard.org/r/5512/