Fix DifFParser preamble parsing with weird unicode.
One of the old review requests on r.rb.o has a diff file which is basically
utf-8 encoded garbage. This was causing an exception because we were trying to
concatenate strings and bytes together, which caused it to try to decode the
utf-8 text as if it were ascii. Adding some 'b' prefixes to our strings keeps
things in the bytes realm.

Testing done:
Looked at the /diff/raw/ page for both the broken review request, as well as a
couple ones that had valid diffs.